backend/requirements.txt
# Django core
django>=5.0,<5.1
djangorestframework>=3.15
djangorestframework-simplejwt[blacklist]>=5.3
drf-spectacular>=0.27
django-filter>=23.5
django-cors-headers>=4.3
django-celery-beat>=2.6
mysqlclient>=2.2

# Server
gunicorn>=21.2
whitenoise>=6.6

# Task queue
celery>=5.3
redis>=5.0

# Security
argon2-cffi>=23.1

# ML — Face Recognition (InsightFace ONNX — no pip package needed, uses buffalo_l models directly)
opencv-python-headless>=4.9
Pillow>=10.2
numpy>=1.26
scikit-learn>=1.4

# Vector DB
chromadb>=0.4.22

# Anti-spoofing
onnxruntime>=1.17

# Utilities
python-dotenv>=1.0
openpyxl>=3.1
django-ratelimit>=4.1
